+++ src/api/authentication/pqsolver.hpp
@@ -9,5 +9,5 @@ public:
 
     ~PQSolver();
 
-    void solvePQ(uint64_t, uint64_t*, uint64_t*);
+    void solvePQ(uint64_t, uint32_t*, uint32_t*);
 };
